daemon/pyproject.toml
[project]
name = "neo-daemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.1",
    "google-generativeai>=0.8.6",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.46.0",
    "websockets>=16.0",
    "httpx>=0.28.1",
    "google-genai>=2.2.0",
    "numpy>=2.4.4",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
]
